+++ drivers/HmIPW-DRAP/driver.ts
@@ -1,0 +1,4 @@
+import { RuntimeBackedDriver } from "../../src/homey/runtime-backed-driver";
+export = class HmIpwDrapDriver extends RuntimeBackedDriver {
+  protected readonly openCcuDriverId = "HmIPW-DRAP";
+};