+++ drivers/HmIPW-DRI16/driver.ts
@@ -1,0 +1,4 @@
+import { RuntimeBackedDriver } from "../../src/homey/runtime-backed-driver";
+export = class HmIpwDri16Driver extends RuntimeBackedDriver {
+  protected readonly openCcuDriverId = "HmIPW-DRI16";
+};