+++ drivers/HmIPW-DRS8/driver.ts
@@ -1,0 +1,4 @@
+import { RuntimeBackedDriver } from "../../src/homey/runtime-backed-driver";
+export = class HmIpwDrs8Driver extends RuntimeBackedDriver {
+  protected readonly openCcuDriverId = "HmIPW-DRS8";
+};